requirements.txt
discord.py==2.4.0
PyYAML==6.0.1
selenium==4.12.0
webdriver-manager==4.0.2
